requirements.txt
gymnasium>=0.29.0
stable-baselines3>=2.3.0
pygame>=2.5.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy>=1.24.0
